Fixed dependency checking bug
Dependencies became stuck on the last build type used, of those available.
For example, if there are 3 build types A, B and C; and the user runs `make A`
and then `make B`, then running `make A` again will not work, because it looks at
the dependencies from B.  The solution was to create folders for each build in the
dependency directory.