Fixed end of range handling
The range end iterator would point to the last element that passed the
range's filter. This is incorrect as standard behaviour is to point
past the end of the container.  Now the end iterator points past
the end of the range as a standard iterator would.  Previously this
error caused a range-based for loop to skip the final valid element.
This symptom no longer occurs.