Renamed equation to polynomial
A new equation concept is slated to be added, so equation was renamed to
polynomial, allowing the name equation to be used again.

Additionally note: using the file name test-polynomial.cc caused a crash in the
program, but renaming that file to test-epolynomial.cc seems to work fine. Some
searching for the root cause appears to narrow the problem to the 6th letter
being the letter p.