Created the event trigger class
The event trigger class is the private piece of the event system, such that an event owner is the only one that can trigger it.  A class using the event trigger can then expose a method to access the trigger's event channel, to distribute to it's clients, which can be subscribed to.